pyproject.toml
[project]
name = "ai-news-brief"
version = "0.1.0"
description = "Automated AI news briefing system using LangGraph and Claude"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.60",
    "langchain-anthropic>=0.3.5",
    "langchain-openai>=0.3.5",
    "langchain-google-genai>=2.0.11",
    "langchain-core>=0.3.30",
    "feedparser>=6.0.11",
    "praw>=7.8.2",
    "python-telegram-bot>=21.11",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.2",
    "arxiv>=2.1.3",
    "requests>=2.32.3",
    "aiohttp>=3.11.11",
    "python-dotenv>=1.0.1",
    "jinja2>=3.1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.8.7",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
